livro

+++ ex003.py
@@ -1,8 +1,12 @@
 # Usaremos um dos 4 tipos primitivos de variaveis para resolver esse exercicio, a variavel int que retorna um numero inteiro.
+
 # Para mostrarmos a soma de maneira mais adequada usaremos as {} seguindo do .format(variavel)
 
-n1 = int(input('Digite o primeiro numero: '))
-n2 = int(input('Digite o segundo numero: '))
+n1 = float(input('Digite o primeiro numero: '))
+n2 = float(input('Digite o segundo numero: '))
+
 s = n1 + n2
 
 print('A soma entre {} e {} é {}'.format(n1, n2, s))
+
+
